appriffy_demo/requirements.txt
asgiref==3.2.3
Django==3.0.2
django-cors-headers==3.2.1
djangorestframework==3.11.0
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.0
